refactor MockClusterInfo into separate header file
