+++ test/mocks/upstream/cluster_info.h
@@ -1,0 +1,41 @@
+#pragma once
+
+#include "envoy/upstream/cluster_manager.h"
+#include "envoy/upstream/upstream.h"
+
+#include "common/stats/stats_impl.h"
+
+#include "test/mocks/runtime/mocks.h"
+#include "test/mocks/stats/mocks.h"
+
+using testing::NiceMock;
+
+namespace Upstream {
+
+class MockClusterInfo : public ClusterInfo {
+public:
+  MockClusterInfo();
+  ~MockClusterInfo();
+
+  // Upstream::ClusterInfo
+  MOCK_CONST_METHOD0(connectTimeout, std::chrono::milliseconds());
+  MOCK_CONST_METHOD0(features, uint64_t());
+  MOCK_CONST_METHOD0(httpCodecOptions, uint64_t());
+  MOCK_CONST_METHOD0(lbType, LoadBalancerType());
+  MOCK_CONST_METHOD0(maintenanceMode, bool());
+  MOCK_CONST_METHOD0(maxRequestsPerConnection, uint64_t());
+  MOCK_CONST_METHOD0(name, const std::string&());
+  MOCK_CONST_METHOD1(resourceManager, ResourceManager&(ResourcePriority priority));
+  MOCK_CONST_METHOD0(sslContext, Ssl::ClientContext*());
+  MOCK_CONST_METHOD0(stats, ClusterStats&());
+  MOCK_CONST_METHOD0(statsScope, Stats::Scope&());
+
+  std::string name_{"fake_cluster"};
+  uint64_t max_requests_per_connection_{};
+  NiceMock<Stats::MockIsolatedStatsStore> stats_store_;
+  ClusterStats stats_;
+  NiceMock<Runtime::MockLoader> runtime_;
+  std::unique_ptr<Upstream::ResourceManager> resource_manager_;
+};
+
+} // Upstream
