+++ test/mocks/upstream/host.h
@@ -3,6 +3,7 @@
 #include "envoy/upstream/upstream.h"
 
 #include "common/stats/stats_impl.h"
+#include "cluster_info.h"
 
 namespace Upstream {
 namespace Outlier {
